requirements.txt
flask==2.3.3
langchain
langchain-community
langchain-openai
chromadb==0.4.24
openai
python-dotenv==1.0.1
requests==2.32.3
numpy==1.26.4
